requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.6.0
attrs==23.2.0
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
Faker==22.6.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pandas==2.1.4
pillow==11.1.0
pluggy==1.3.0
pydantic==2.6.0
pydantic_core==2.16.1
pydentic==0.0.1.dev3
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.1.0
